pipeline-engine/requirements.txt
# YARUKSAİ Pipeline Engine — Dependencies
# ─────────────────────────────────────────
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0,<1.0
pydantic>=2.0,<3.0
python-dotenv>=1.0,<2.0

# CrewAI + LLM providers
crewai>=0.86.0,<1.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.3.0

# SaaS Layer — Auth, PDF, Billing
python-jose[cryptography]==3.3.0
cryptography>=3.4.8,<42.0.0
PyJWT>=2.8.0,<3.0
reportlab>=4.0,<5.0
qrcode[pil]>=7.4
